// frontend/src/api/model/detectorLogbook.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */
import type { UserSummary } from './userSummary';
import type { EntryTypeEnum } from './entryTypeEnum';
import type { SourceEnum } from './sourceEnum';

export interface DetectorLogbook {
  readonly id: string;
  readonly author: UserSummary;
  readonly modified_by: UserSummary;
  is_deleted?: boolean;
  /** @nullable */
  deleted_at?: string | null;
  readonly created: string;
  /** @nullable */
  readonly modified: string | null;
  /** Detailed description of activity made on the detector. */
  text: string;
  /** Private logbook will be visible for maintainers of detector and for dosportal admins. */
  public?: boolean;
  /** Category of the logbook entry.

* `reset` - Reset
* `sync` - Sync
* `maintenance` - Maintenance
* `note` - Note
* `location_update` - Location update
* `calibration` - Calibration
* `other` - Other */
  entry_type?: EntryTypeEnum;
  /** Origin of the logbook entry.

* `web` - Web
* `api` - API
* `qr` - QR
* `auto` - Automatic
* `other` - Other */
  source?: SourceEnum;
  /**
   * GPS latitude of the location
   * @nullable
   */
  latitude?: number | null;
  /**
   * GPS longitude of the location
   * @nullable
   */
  longitude?: number | null;
  /**
   * Altitude of the location in meters
   * @nullable
   */
  altitude?: number | null;
  /**
   * Location description in text format (e.g., address)
   * @maxLength 255
   * @nullable
   */
  location_text?: string | null;
  /** @nullable */
  deleted_by?: number | null;
  detector: string;
}